# sql: replace `walkPlan` in `startExec`