`startExec` now traverses a `planNode` tree using the `Input` and
`InputCount` methods instead of `walkPlan`.

Epic: None

Release note: None