@mgartner 2025-01-08
```
name                                        old time/op    new time/op    delta
EndToEnd/kv-read/vectorize=on/Prepared-16      145µs ± 1%     139µs ± 1%  -4.15%  (p=0.000 n=20+20)
EndToEnd/kv-read/vectorize=off/Prepared-16     139µs ± 1%     133µs ± 1%  -4.00%  (p=0.000 n=20+17)
EndToEnd/kv-read/vectorize=on/Simple-16        212µs ± 2%     207µs ± 1%  -2.61%  (p=0.000 n=20+19)
EndToEnd/kv-read/vectorize=off/Simple-16       206µs ± 1%     201µs ± 2%  -2.48%  (p=0.000 n=20+20)

name                                        old alloc/op   new alloc/op   delta
EndToEnd/kv-read/vectorize=on/Prepared-16     20.5kB ± 0%    20.5kB ± 0%  -0.29%  (p=0.000 n=20+19)
EndToEnd/kv-read/vectorize=off/Prepared-16    21.8kB ± 0%    21.7kB ± 0%  -0.22%  (p=0.000 n=19+20)
EndToEnd/kv-read/vectorize=on/Simple-16       33.4kB ± 0%    33.4kB ± 0%    ~     (p=0.815 n=20+20)
EndToEnd/kv-read/vectorize=off/Simple-16      33.8kB ± 0%    33.8kB ± 0%    ~     (p=0.784 n=20+20)

name                                        old allocs/op  new allocs/op  delta
EndToEnd/kv-read/vectorize=off/Prepared-16       179 ± 0%       178 ± 0%  -0.31%  (p=0.001 n=20+16)
EndToEnd/kv-read/vectorize=on/Simple-16          279 ± 0%       279 ± 0%  -0.14%  (p=0.001 n=20+16)
EndToEnd/kv-read/vectorize=on/Prepared-16        186 ± 0%       186 ± 0%    ~     (p=0.614 n=20+19)
EndToEnd/kv-read/vectorize=off/Simple-16         267 ± 0%       267 ± 0%    ~     (all equal)
```