@mgartner 2025-01-08
I get a bizarre inconsistent result with the sysbench microbenchmark:

```
name                                        old time/op    new time/op    delta
Sysbench/SQL/1node_local/oltp_read_only-16    2.64ms ± 1%    2.72ms ± 2%  +3.39%  (p=0.000 n=20+20)

name                                        old alloc/op   new alloc/op   delta
Sysbench/SQL/1node_local/oltp_read_only-16     835kB ± 0%     836kB ± 0%    ~     (p=0.799 n=20+20)

name                                        old allocs/op  new allocs/op  delta
Sysbench/SQL/1node_local/oltp_read_only-16     3.90k ± 1%     3.90k ± 0%    ~     (p=0.703 n=20+20)
```