@mgartner 2025-01-08
Reran some more end-to-end benchmarks and oddly the benefit is not seen:

```
name                                                                        old time/op    new time/op    delta
EndToEnd/tpcc-stock-level/vectorize=on/Prepared-16                             265µs ± 1%     263µs ± 2%  -0.71%  (p=0.020 n=20+20)
EndToEnd/json-comp-insert/vectorize=on/Prepared-16                             272µs ± 3%     270µs ± 3%  -0.71%  (p=0.046 n=20+20)
EndToEnd/tpcc-delivery/vectorize=on/Prepared-16                                217µs ± 1%     216µs ± 1%  -0.54%  (p=0.001 n=19+19)
EndToEnd/single-col-histogram-bounded-range-big/vectorize=on/Prepared-16       207µs ± 2%     206µs ± 1%  -0.51%  (p=0.020 n=20+20)
EndToEnd/kv-read-const/vectorize=on/Prepared-16                                145µs ± 1%     144µs ± 1%  -0.45%  (p=0.002 n=16+20)
EndToEnd/kv-read/vectorize=on/Prepared-16                                      146µs ± 1%     145µs ± 1%  -0.40%  (p=0.022 n=19+20)
EndToEnd/many-columns-and-indexes-b/vectorize=on/Prepared-16                   167µs ± 1%     167µs ± 1%  -0.29%  (p=0.036 n=19+18)
EndToEnd/comp-insert-on-conflict/vectorize=on/Prepared-16                      398µs ± 1%     397µs ± 1%  -0.26%  (p=0.026 n=19+18)
EndToEnd/tpcc-new-order/vectorize=on/Prepared-16                               151µs ± 2%     150µs ± 1%    ~     (p=0.155 n=20+20)
EndToEnd/many-columns-and-indexes-a/vectorize=on/Prepared-16                   163µs ± 1%     163µs ± 1%    ~     (p=0.687 n=19+20)
EndToEnd/many-columns-and-indexes-c/vectorize=on/Prepared-16                   196ms ± 1%     196ms ± 1%    ~     (p=0.086 n=20+20)
EndToEnd/many-columns-and-indexes-d/vectorize=on/Prepared-16                  15.9ms ± 1%    16.0ms ± 1%    ~     (p=0.201 n=20+17)
EndToEnd/comp-pk/vectorize=on/Prepared-16                                      204µs ± 2%     203µs ± 2%    ~     (p=0.089 n=19+20)
EndToEnd/single-col-histogram-range/vectorize=on/Prepared-16                   174µs ± 1%     173µs ± 1%    ~     (p=0.063 n=20+20)
EndToEnd/single-col-histogram-bounded-range-small/vectorize=on/Prepared-16     199µs ± 1%     198µs ± 2%    ~     (p=0.072 n=20+20)
EndToEnd/json-insert/vectorize=on/Prepared-16                                 30.3ms ± 1%    30.3ms ± 0%    ~     (p=0.214 n=19+20)
EndToEnd/batch-insert-one/vectorize=on/Prepared-16                            30.6ms ± 1%    30.6ms ± 1%    ~     (p=0.989 n=19+20)
EndToEnd/batch-insert-many/vectorize=on/Prepared-16                           35.2ms ± 1%    35.2ms ± 1%    ~     (p=0.242 n=20+20)
EndToEnd/const-agg/vectorize=on/Prepared-16                                    205µs ± 2%     205µs ± 2%    ~     (p=0.429 n=20+20)
EndToEnd/ored-preds-100/vectorize=on/Prepared-16                               284µs ± 3%     284µs ± 3%    ~     (p=0.640 n=20+20)
EndToEnd/ored-preds-using-params-100/vectorize=on/Prepared-16                 1.02ms ± 2%    1.02ms ± 1%    ~     (p=0.835 n=19+20)

name                                                                        old alloc/op   new alloc/op   delta
EndToEnd/single-col-histogram-bounded-range-big/vectorize=on/Prepared-16      76.2kB ± 1%    74.7kB ± 4%  -1.99%  (p=0.016 n=18+20)
EndToEnd/json-insert/vectorize=on/Prepared-16                                 5.69MB ± 1%    5.67MB ± 1%  -0.22%  (p=0.049 n=17+17)
EndToEnd/single-col-histogram-range/vectorize=on/Prepared-16                  36.0kB ± 0%    35.9kB ± 0%  -0.14%  (p=0.001 n=20+18)
EndToEnd/kv-read/vectorize=on/Prepared-16                                     21.3kB ± 0%    21.3kB ± 0%    ~     (p=0.869 n=19+19)
EndToEnd/kv-read-const/vectorize=on/Prepared-16                               20.2kB ± 0%    20.2kB ± 0%    ~     (p=0.638 n=18+20)
EndToEnd/tpcc-new-order/vectorize=on/Prepared-16                              22.1kB ± 0%    22.1kB ± 0%    ~     (p=0.161 n=19+18)
EndToEnd/tpcc-delivery/vectorize=on/Prepared-16                               36.3kB ± 0%    36.3kB ± 0%    ~     (p=0.076 n=18+19)
EndToEnd/tpcc-stock-level/vectorize=on/Prepared-16                            89.0kB ± 0%    88.0kB ± 4%    ~     (p=0.578 n=18+20)
EndToEnd/many-columns-and-indexes-a/vectorize=on/Prepared-16                  25.1kB ± 0%    25.1kB ± 0%    ~     (p=0.598 n=20+18)
EndToEnd/many-columns-and-indexes-b/vectorize=on/Prepared-16                  30.2kB ± 1%    29.1kB ± 9%    ~     (p=0.873 n=19+20)
EndToEnd/many-columns-and-indexes-c/vectorize=on/Prepared-16                   125MB ± 0%     125MB ± 0%    ~     (p=0.076 n=17+19)
EndToEnd/many-columns-and-indexes-d/vectorize=on/Prepared-16                  11.7MB ± 0%    11.7MB ± 0%    ~     (p=0.729 n=19+19)
EndToEnd/comp-pk/vectorize=on/Prepared-16                                     56.7kB ± 0%    56.9kB ± 1%    ~     (p=0.945 n=19+20)
EndToEnd/comp-insert-on-conflict/vectorize=on/Prepared-16                      114kB ± 0%     113kB ± 1%    ~     (p=0.053 n=17+20)
EndToEnd/single-col-histogram-bounded-range-small/vectorize=on/Prepared-16    29.1kB ± 0%    30.1kB ±11%    ~     (p=0.390 n=16+20)
EndToEnd/json-comp-insert/vectorize=on/Prepared-16                            54.1kB ± 1%    54.0kB ± 1%    ~     (p=0.209 n=18+20)
EndToEnd/batch-insert-one/vectorize=on/Prepared-16                            6.50MB ± 3%    6.56MB ± 1%    ~     (p=0.380 n=20+19)
EndToEnd/batch-insert-many/vectorize=on/Prepared-16                           10.9MB ± 3%    11.0MB ± 3%    ~     (p=0.270 n=20+19)
EndToEnd/const-agg/vectorize=on/Prepared-16                                   67.0kB ± 1%    66.9kB ± 1%    ~     (p=0.747 n=18+19)
EndToEnd/ored-preds-100/vectorize=on/Prepared-16                               171kB ± 2%     171kB ± 1%    ~     (p=0.560 n=20+16)
EndToEnd/ored-preds-using-params-100/vectorize=on/Prepared-16                  946kB ± 4%     939kB ± 2%    ~     (p=0.311 n=19+19)

name                                                                        old allocs/op  new allocs/op  delta
EndToEnd/single-col-histogram-bounded-range-big/vectorize=on/Prepared-16         296 ± 1%       285 ± 7%  -3.78%  (p=0.000 n=18+20)
EndToEnd/single-col-histogram-range/vectorize=on/Prepared-16                     280 ± 0%       279 ± 0%  -0.40%  (p=0.000 n=20+19)
EndToEnd/json-comp-insert/vectorize=on/Prepared-16                               445 ± 0%       444 ± 0%  -0.38%  (p=0.000 n=19+19)
EndToEnd/comp-pk/vectorize=on/Prepared-16                                        464 ± 0%       463 ± 0%  -0.22%  (p=0.000 n=15+15)
EndToEnd/comp-insert-on-conflict/vectorize=on/Prepared-16                        939 ± 0%       937 ± 0%  -0.21%  (p=0.000 n=20+20)
EndToEnd/json-insert/vectorize=on/Prepared-16                                  35.3k ± 0%     35.3k ± 0%  -0.12%  (p=0.001 n=20+18)
EndToEnd/kv-read/vectorize=on/Prepared-16                                        189 ± 0%       189 ± 0%    ~     (p=1.000 n=20+20)
EndToEnd/tpcc-new-order/vectorize=on/Prepared-16                                 203 ± 0%       203 ± 0%    ~     (all equal)
EndToEnd/tpcc-delivery/vectorize=on/Prepared-16                                  293 ± 0%       293 ± 0%    ~     (p=0.635 n=18+19)
EndToEnd/tpcc-stock-level/vectorize=on/Prepared-16                               730 ± 0%       722 ± 3%    ~     (p=0.417 n=19+20)
EndToEnd/many-columns-and-indexes-a/vectorize=on/Prepared-16                     241 ± 0%       241 ± 0%    ~     (p=0.071 n=17+20)
EndToEnd/many-columns-and-indexes-b/vectorize=on/Prepared-16                     276 ± 0%       269 ± 6%    ~     (p=0.067 n=18+20)
EndToEnd/many-columns-and-indexes-c/vectorize=on/Prepared-16                    896k ± 0%      898k ± 0%    ~     (p=0.064 n=17+19)
EndToEnd/many-columns-and-indexes-d/vectorize=on/Prepared-16                    218k ± 0%      218k ± 0%    ~     (p=0.485 n=17+19)
EndToEnd/single-col-histogram-bounded-range-small/vectorize=on/Prepared-16       248 ± 0%       255 ± 9%    ~     (p=0.274 n=16+20)
EndToEnd/batch-insert-one/vectorize=on/Prepared-16                             38.3k ± 1%     38.2k ± 1%    ~     (p=0.087 n=20+20)
EndToEnd/batch-insert-many/vectorize=on/Prepared-16                            59.2k ± 0%     59.2k ± 1%    ~     (p=0.130 n=15+17)
EndToEnd/const-agg/vectorize=on/Prepared-16                                      454 ± 1%       453 ± 1%    ~     (p=0.252 n=17+18)
EndToEnd/ored-preds-100/vectorize=on/Prepared-16                               2.03k ± 0%     2.03k ± 0%    ~     (p=0.262 n=20+16)
EndToEnd/ored-preds-using-params-100/vectorize=on/Prepared-16                  5.13k ± 4%     5.12k ± 6%    ~     (p=0.572 n=19+20)
EndToEnd/kv-read-const/vectorize=on/Prepared-16                                  181 ± 0%       181 ± 0%  +0.19%  (p=0.019 n=16+20)
```

Sysbench still shows a slowdown, though `oltp_read_write` is less pronounced: 

```
name                                         old time/op    new time/op    delta
Sysbench/SQL/1node_local/oltp_read_write-16    4.61ms ± 2%    4.68ms ± 1%  +1.58%  (p=0.000 n=20+20)
Sysbench/SQL/1node_local/oltp_read_only-16     2.64ms ± 1%    2.73ms ± 1%  +3.20%  (p=0.000 n=19+19)

name                                         old alloc/op   new alloc/op   delta
Sysbench/SQL/1node_local/oltp_read_write-16    1.32MB ± 9%    1.30MB ±10%    ~     (p=0.968 n=20+20)
Sysbench/SQL/1node_local/oltp_read_only-16      836kB ± 0%     836kB ± 0%  +0.09%  (p=0.038 n=20+19)

name                                         old allocs/op  new allocs/op  delta
Sysbench/SQL/1node_local/oltp_read_write-16     6.73k ± 1%     6.71k ± 1%  -0.22%  (p=0.032 n=18+20)
Sysbench/SQL/1node_local/oltp_read_only-16      3.90k ± 0%     3.90k ± 1%    ~     (p=0.482 n=20+20)
```