@mgartner 2025-01-08
I re-ran the sysbench test with `sql.metrics.statement_details.enabled = false` and `sql.metrics.transaction_details.enabled = false` and measured a nice speedup:

```
name                                         old time/op    new time/op    delta
Sysbench/SQL/1node_local/oltp_read_only-16     2.67ms ± 1%    2.58ms ± 2%  -3.37%  (p=0.000 n=20+20)
Sysbench/SQL/1node_local/oltp_read_write-16    4.63ms ± 1%    4.52ms ± 2%  -2.29%  (p=0.000 n=20+20)

name                                         old alloc/op   new alloc/op   delta
Sysbench/SQL/1node_local/oltp_read_write-16    1.36MB ± 0%    1.28MB ±11%  -5.71%  (p=0.001 n=16+20)
Sysbench/SQL/1node_local/oltp_read_only-16      804kB ± 0%     803kB ± 0%  -0.11%  (p=0.032 n=20+20)
```

name                                         old allocs/op  new allocs/op  delta
Sysbench/SQL/1node_local/oltp_read_only-16      3.78k ± 0%     3.78k ± 0%    ~     (p=0.243 n=20+19)
Sysbench/SQL/1node_local/oltp_read_write-16     6.50k ± 1%     6.50k ± 0%    ~     (p=0.602 n=19+20)